+++ CHANGELOG.md
@@ -20,7 +20,11 @@ This document details all notable changes to the WSU HRS Child Theme. Uses [Sema
 ### Removed (for deprecated features removed in this release)
 -->
 
-## 0.20.0~build-odbc-connections (unreleased)
+## 0.20.1~build-odbc-connections (unreleased)
+
+### Fixed
+
+- Match ER Awards template to default page template and fix phpcs issues.
 
 ### Changed